requirements.txt
google-cloud-documentai==2.15.0
google-api-core==2.11.1
google-cloud-core==2.3.3
protobuf==4.24.4
grpcio==1.59.0
googleapis-common-protos==1.60.0 